refactor: :bulb: separate constants files
