+++ src/const/about.ts
@@ -1,92 +1,9 @@
-import type { FC } from 'react';
-
-import arcticCodeVault from '../public/images/achievements/arctic-code-vault-contributor-default.png';
-import pullShark from '../public/images/achievements/pull-shark-default.png';
-import quickdraw from '../public/images/achievements/quickdraw-default--light.png';
-import yolo from '../public/images/achievements/yolo-default.png';
-
-import FacebookIcon from './components/Icons/FacebookIcon/FacebookIcon';
-import GithubIcon from './components/Icons/GithubIcon/GithubIcon';
-import InstagramIcon from './components/Icons/InstagramIcon/InstagramIcon';
-import ThreadsIcon from './components/Icons/ThreadsIcon/ThreadsIcon';
+import arcticCodeVault from '../../public/images/achievements/arctic-code-vault-contributor-default.png';
+import pullShark from '../../public/images/achievements/pull-shark-default.png';
+import quickdraw from '../../public/images/achievements/quickdraw-default--light.png';
+import yolo from '../../public/images/achievements/yolo-default.png';
 
 import type { StaticImageData } from 'next/image';
-
-type NavContents = {
-  id: number;
-  href: string;
-  title: string;
-};
-
-// ナビゲーションコンテンツ
-export const NAV_CONTENTS: NavContents[] = [
-  {
-    id: 1,
-    href: '/',
-    title: 'Home',
-  },
-  {
-    id: 2,
-    href: '/about',
-    title: 'About',
-  },
-  {
-    id: 3,
-    href: '/projects',
-    title: 'Projects',
-  },
-  {
-    id: 4,
-    href: '/articles',
-    title: 'Articles',
-  },
-];
-
-type SocialMedia = {
-  id: number;
-  href: string;
-  describedby: string;
-  text: string;
-  component: FC;
-};
-
-// ソーシャルメディアアイコンのリスト
-export const SOCIAL_MEDIA_LIST: SocialMedia[] = [
-  {
-    id: 1,
-    href: 'https://github.com/Bear29ers',
-    describedby: 'navlink-id-1',
-    text: 'github',
-    component: GithubIcon,
-  },
-  {
-    id: 2,
-    href: 'https://www.instagram.com/bear_27earl',
-    describedby: 'navlink-id-2',
-    text: 'instagram',
-    component: InstagramIcon,
-  },
-  {
-    id: 3,
-    href: 'https://www.threads.net/@bear_27earl',
-    describedby: 'navlink-id-3',
-    text: 'threads',
-    component: ThreadsIcon,
-  },
-  {
-    id: 4,
-    href: 'https://www.facebook.com/kumakuma1129',
-    describedby: 'navlink-id-4',
-    text: 'facebook',
-    component: FacebookIcon,
-  },
-];
-
-// Homeページテキスト
-export const HOME_TEXT = {
-  title: 'Code and Design make vision a reality',
-  description: `As a front-end developer and designer, I really enjoy giving shape to my own ideas.\nExplore my personality and latest projests and articles in web development.`,
-};
 
 // Aboutページテキスト
 export const PROFILE_TEXT = {
