+++ src/const/home.ts
@@ -1,0 +1,5 @@
+// Homeページテキスト
+export const HOME_TEXT = {
+  title: 'Code and Design make vision a reality',
+  description: `As a front-end developer and designer, I really enjoy giving shape to my own ideas.\nExplore my personality and latest projests and articles in web development.`,
+};
