+++ src/const/nav.ts
@@ -1,0 +1,29 @@
+type NavContents = {
+  id: number;
+  href: string;
+  title: string;
+};
+
+// ナビゲーションコンテンツ
+export const NAV_CONTENTS: NavContents[] = [
+  {
+    id: 1,
+    href: '/',
+    title: 'Home',
+  },
+  {
+    id: 2,
+    href: '/about',
+    title: 'About',
+  },
+  {
+    id: 3,
+    href: '/projects',
+    title: 'Projects',
+  },
+  {
+    id: 4,
+    href: '/articles',
+    title: 'Articles',
+  },
+];
